+++ NEWS.md
@@ -20,8 +20,10 @@
   up component evaluation and linearisation. (version `2.13.0.9006`)
 * Allow index extraction from duplicated tags in `bru_index()`, returning
   all matching indices. (version `2.13.0.9014`)
-* Code refactor and storage streamlining for `bru_obs()` to prepare for
-  future extensions. (version `2.13.0.9015`)
+* Code refactor and storage updates for `bru_obs()` and `bru_comp()` objects
+  to prepare for future extensions. For 2.14, the compatibility option
+  `bru_compat_pre_2_14_enable` is set to `TRUE` (version `2.13.0.9015`,
+  `2.13.0.9017`, `2.13.0.9018`)
 * Add `block_response` element to `bru_obs(aggregate_input)` list input,
   to allow specifying a block-wise response variable to match for aggregated
   predictors. (version `2.13.0.9016`)